Move existing file path implementation under FilePath abstraction to make file-specific semantics explicit
